superjob-db-migrations/requirements.txt
# superjob-db-migrations/requirements.txt
alembic>=1.13.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary